requirements.txt
aiohttp==3.9.5
aiohttp-cors==0.7.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorful==0.5.6
contourpy==1.2.1
cycler==0.12.1
distlib==0.3.8
dnspython==2.6.1
email_validator==2.1.2
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.1
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
google-api-core==2.19.0
google-auth==2.30.0
googleapis-common-protos==1.63.1
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdit-py-plugins==0.4.1
mdurl==0.1.2
memray==1.12.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.2.1
numpy==2.0.0
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
opencensus==0.11.4
opencensus-context==0.1.3
orjson==3.10.5
packaging==24.1
pandas==2.2.2
pillow==10.2.0
platformdirs==4.2.2
prometheus_client==0.20.0
proto-plus==1.23.0
protobuf==4.25.3
py-spy==0.3.14
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
ray==2.24.0
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
starlette==0.37.2
sympy==1.12
tensorboardX==2.6.2.2
textual==0.69.0
torch==2.3.1+cu118
torchaudio==2.3.1+cu118
torchvision==0.18.1+cu118
tqdm==4.66.4
triton==2.3.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
virtualenv==20.26.2
watchfiles==0.22.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
